Split Program.cs option plumbing into Program.Options.cs
Move the command-line option definitions (aliases, Option fields, config
help rows, and the Get*Option/GetRootCommand factories) into a dedicated
partial class file. This keeps the argument-parsing behavior (Main, Run,
ParseKeyValuePairs) separate from the option wiring and makes the CLI
easier to navigate.
